Adding filter function Adding swiftlint Tidying code indentation and whitespace
