+++ .swiftlint.yml
@@ -1,0 +1,53 @@
+opt_in_rules:
+  - closure_end_indentation
+  - closure_spacing
+  - empty_count
+  - fatal_error_message
+  - first_where
+  - joined_default_parameter
+  - literal_expression_end_indentation
+  - multiline_arguments
+  - multiline_parameters
+  - operator_usage_whitespace
+  - overridden_super_call
+  - pattern_matching_keywords
+  - prohibited_super_call
+  - redundant_nil_coalescing
+  - sorted_first_last
+  - switch_case_on_newline
+  - trailing_closure
+  - unneeded_parentheses_in_closure_argument
+  - vertical_parameter_alignment_on_call
+  - yoda_condition
+  - file_name
+  - function_default_parameter_at_end
+  - fallthrough
+  - array_init
+  - contains_over_first_not_nil
+
+disabled_rules:
+  - function_body_length
+  - file_length
+  - type_body_length
+  - function_parameter_count
+  - nested_types
+
+
+excluded:
+  - Carthage
+  - fastlane
+  - CoreTests
+  - AmbidexterTests
+  - Ambidexter.playground
+
+identifier_name:
+  excluded:
+    - id
+    - on
+    - to
+    - qr
+    - x
+    - y
+
+file_name: 
+  excluded: ["HttpData+Payments.swift"]
