+++ Sources/Bind/Relay.swift
@@ -1,7 +1,5 @@
-import Foundation
-
 public final class Relay: Output<Void> {
-  public func fire() {
-    update(withValue: ())
-  }
+    public func fire() {
+        update(withValue: ())
+    }
 }
