+++ Tests/BindTests/Output+Extension.swift
@@ -1,13 +1,13 @@
 @testable import Bind
 
 extension Output {
-  var latest: Value? {
-    var value: Value?
-    
-    bind { current in
-      value = current
+    var latest: Value? {
+        var value: Value?
+
+        bind { current in
+            value = current
+        }
+
+        return value
     }
-    
-    return value
-  }
 }
