+++ Tests/BindTests/PrinterMock.swift
@@ -1,8 +1,8 @@
 @testable import Bind
 
 final class PrinterMock: Printable {
-  var printValues: [String] = []
-  func print(_ value: String) {
-    printValues.append(value)
-  }
+    var printValues: [String] = []
+    func print(_ value: String) {
+        printValues.append(value)
+    }
 }
