+++ Tests/BindTests/XCTestManifests.swift
@@ -3,7 +3,7 @@ import XCTest
 #if !canImport(ObjectiveC)
 public func allTests() -> [XCTestCaseEntry] {
     return [
-        testCase(TriggerTests.allTests),
+        testCase(TriggerTests.allTests)
     ]
 }
 #endif
